requirements/main.txt
gym>=0.24.1
hydra-core>=1.2.0
matplotlib>=3.5.2
numpy>=1.22.4
torch>=1.11.0
torchdyn>=1.0.3
termcolor>=1.1.0
imageio>=2.19.0